Modularise the scoring function
