+++ src/js/score/apiDetails.js
@@ -1,0 +1,6 @@
+const api = {
+  url: 'https://us-central1-js-capstone-backend.cloudfunctions.net/api',
+  id: 'Qa2XUOTtOtFp4SJONyU5',
+};
+
+export default api;
